requirements.txt
appdirs==1.4.4
asgiref==3.2.7
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
cymem==2.0.3
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.5
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-summernote==0.8.11.6
filelock==3.0.12
idna==2.9
murmurhash==1.0.2
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.3
Pillow==7.1.2
plac==1.1.3
preshed==3.0.2
psycopg2-binary==2.8.5
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
spacy==2.2.4
sqlparse==0.3.1
srsly==1.0.2
thinc==7.4.0
tqdm==4.46.0
tweepy==3.8.0
urllib3==1.25.9
virtualenv==20.0.21
wasabi==0.6.0
whitenoise==5.1.0
gunicorn
